Removed lots of useless profiles and parametrized toolchain p2 repo
To build scalariform with Maven, you need to provide a p2 repository that
contains the Scala compiler + library. For instance, you can use any of the
Scala IDE update-site. When building, do so by providing the update-site URL
value via the Maven -Drepo.scala-ide.toolchain argument.

The Maven build is used only for producing a P2 repository for scalariform.  If
you want to build scalariform as a library, please use the Sbt build.